// src/types/enums/battleState.ts
export enum BattleState {
  NOT_STARTED = "NOT_STARTED",
  BATTLE = "BATTLE",
  BATTLE_ENDED = "BATTLE_ENDED",
  BATTLE_START = "BATTLE_START",
  USER_VICTORY = "USER_VICTORY",
  USER_SWITCHING_POKEMON = "USER_SWITCHING_POKEMON",
  USER_POKEMON_FAINTED = "USER_POKEMON_FAINTED",
  USER_ATTACKING = "USER_ATTACKING",
  OPPONENT_VICTORY = "OPPONENT_VICTORY",
  OPPONENT_SWITCHING_POKEMON = "OPPONENT_SWITCHING_POKEMON",
  OPPONENT_POKEMON_FAINTED = "OPPONENT_POKEMON_FAINTED",
  OPPONENT_ATTACKING = "OPPONENT_ATTACKING",
  CANNOT_START = "CANNOT_START",
  BATTLE_STOPPED = "BATTLE_STOPPED",
  // WILD POKEMON MODE ONLY
  CATCHING = "CATCHING",
  CAUGHT = "CAUGHT",
  ROUND_DONE = "ROUND_DONE",
  WILD_POKEMON_FAINTED = "WILD_POKEMON_FAINTED",
  WILD_POKEMON_APPEAR = "WILD_POKEMON_APPEAR",
  LAST_ROUND_DONE = "LAST_ROUND_DONE",
}